Extend expand_relativepathes function
It is now possible to:
a) Use lists as begin- and endtag. Instead of implementing several runthroughs with the same attributes and different begin/endtags this should be faster to calculate and easier to implement.

b) It is possible to use wildcard "*" at the end of an attribute, e.g. "se_*". If an attribute starts with "se_" the value will get parsed and relative item definitions are expaneded.