improve windows compatability with update of file operations to explicit usage of utf-8
Windows will not return utf-8 as preferred encoding but rather something like cp1252, which is bad for SHNG's UTF-8 files.
This fixes read and write operations for text files (logics, yaml) to explicit use utf-8 encoding.
Since also the paths separator is '\' and not '/' the standard lib pathlib is introduced to make the handling of paths and files more portable.